+++ NEWS.md
@@ -19,6 +19,8 @@
   sessions.
 - Add `predict` method to obtain conditional mean estimates and prediction intervals, both with or without previous observations.
 - Use automatic differentiation to calculate Satterthwaite adjusted degrees of freedom calculation, resulting in 10-fold speed-up of the Satterthwaite calculations after the initial model fit.
+- Two new methods added for mmrm models: `model.matrix()` and `terms()`.
+
 
 ### Miscellaneous
 
@@ -40,6 +42,9 @@
   This is now fixed.
 - Previously `mmrm` can provide non-finite values in Jacobian calculations.
   This will raise an error now.
+- The `model.frame()` method has been updated to return a data frame the size of 
+  the number of observations utilized in the model for all combinations of the 
+  `model.frame(include)` argument.
 
 # mmrm 0.2.2